[whywhathow]  fix: 修改数据库表字段 导致 buildTree 方法失效 reason: db 字段 由 int ->bigInt , service 方面 没有进行对应的映射, 导致TreeUtil.build() 方法失效.

+++ fun-api/fun-system-api/src/main/java/com/fun/system/api/mapper/MenuMapper.java
@@ -42,6 +42,6 @@ public interface MenuMapper extends BaseMapper<Menu> {
 
     List<Menu> selectByRoleId(Long roleId);
 
-    Boolean updateByMenuId(@Param("menuId") Integer menuId);
+    Boolean updateByMenuId(@Param("menuId") Long menuId);
 }
